% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covNB.R
\name{covNB}
\alias{covNB}
\title{Computes the covariance between net benefit values from 2 distinct datasets.}
\usage{
covNB(data1 = NULL, data2 = NULL, will2pay = NULL, extraOutput = FALSE,
  CI = 0.95, treatResponse = "beneficial")
}
\arguments{
\item{data1}{a dataset with 4 columns holding respectively, ID, treatment outcome,
treatment cost and treatment arm.}

\item{data2}{a dataset with 4 columns holding respectively, ID, treatment outcome,
treatment cost and treatment arm.}

\item{will2pay}{a numeric vector of willingness-to-pay thresholds.}

\item{extraOutput}{a boolean set to FALSE by default, if set to TRUE other information are returned
in addition to \code{covNB}, the covariance between net benefits from the two datasets.}

\item{CI}{confidence interval.}

\item{treatResponse}{a character, default is \code{beneficial} i.e. the treatment resulted in beneficial response;
otherwise \code{harmful}, the treatement resulted in harmful outcome}
}
\value{
a list which holds the below items; items other than \code{covNB} are optional, set the 
the parameter \code{extraOutput} to have those items returned:
\code{covNB} covariance between net benefits from the two datasets
\code{rhoNB} correlation between net benefits from the two datasets
\code{covCostA} covariance of costs between the two datasets for the 1st treatment arm
\code{covCostB} covariance of costs between the two datasets for the 2nd treatment arm
\code{covOutcomeA} covariance of outcomes between the two datasets for the 1st treatment arm
\code{covOutcomeB} covariance of outcomes between the two datasets for the 2nd treatment arm
}
\description{
The function takes 2 datasets and a vector of wilingness-to-pay values.
The dataset MUST have 4 columns \code{ID}, \code{RESPONSE}, \code{COST} and \code{TREATMENT}
which hold respectively unique subject identifiers, treatment outcome, treatment cost
and 2 treatment arms names. If no values are for wilingness-to-pay, 101 default values ranging
from 0 to 500,000 are used to compute net benefits and covariance between net benefits from the 2 datasets
}
\details{
to be written
}
\examples{
{

# load examples datasets
data(dataset1); data(dataset2)

# covNB computation using the default willingness-to-pay thresholds
covValues <- covNB(data1=dataset1, data2=dataset2)

# covNB computation using the default willingness-to-pay thresholds 
#and request extra information.
covValues <- covNB(data1=dataset1, data2=dataset2, extraOutput=TRUE)

}

}
\author{
Amadou Gaye & Felix Achana
}

