\name{dataset1}
\alias{dataset1}
\docType{data}
\title{Dataset to investigate cost-effectiveness for}
\description{A dataset with 4 columns \code{ID}, \code{RESPONSE}, \code{COST} and \code{TREATMENT}}
\usage{data(dataset1)}
\format{
  A table of 4 columns
  \describe{
    \item{\code{ID}}{Unique identifier for subjects}
    \item{\code{RESPONSE}}{Outcome of the treatment}
    \item{\code{COST}}{Cost of the treatment}
    \item{\code{Treatment}}{Treatment arm identifiers}
  }
}


\examples{data(dataset1)}
