% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffNB.R
\name{diffNB}
\alias{diffNB}
\title{Computes difference between net benefits from 2 distinct datasets.}
\usage{
diffNB(data1 = NULL, data2 = NULL, will2pay = NULL)
}
\arguments{
\item{data1}{a dataset with 4 columns holding respectively, ID, treatment outcome,
treatment cost and treatment arm.}

\item{data2}{a dataset with 4 columns holding respectively, ID, treatment outcome,
treatment cost and treatment arm.}

\item{will2pay}{a numeric vector of willingness-to-pay thresholds.}
}
\value{
a list which holds the below items; items other than \code{covNB} are optional, set the 
the parameter \code{extraOutput} to have those items returned:
\code{diffNB} to be written
\code{sediffNB} to be written
\code{lcldiffNB} to be written
\code{ucldiffNB} to be written
\code{pvaluediffNB} to be written
}
\description{
The function takes 2 datasets and a vector of wilingness-to-pay values.
The dataset MUST have 4 columns \code{ID}, \code{RESPONSE}, \code{COST} and \code{TREATMENT}
which hold respectively unique subject identifiers, treatment outcome, treatment cost
and 2 treatment arms names. If no values are for wilingness-to-pay, 101 default values ranging
from 0 to 500,000 are used to estimate the difference between net benefits from the 2 datasets
}
\details{
to be written
}
\examples{
{

# load examples datasets
data(dataset1); data(dataset2)

# diffNB computation using the default willingness-to-pay thresholds
results <- diffNB(data1=dataset1, data2=dataset2)

}

}
\author{
Amadou Gaye & Felix Achana
}

