% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaries.R
\name{getSummaries}
\alias{getSummaries}
\title{Generates various information required within different functions}
\usage{
getSummaries(dt)
}
\arguments{
\item{data}{a dataset with 4 columns holding respectively, ID, treatment outcome,
treatment cost and treatment arm.}
}
\value{
a list with the following items:
\code{idxA} indices of the observation in the first treatment arm
\code{idxB} indices of the observation in the second treatment arm
\code{cost} cost values
\code{meanCostA} mean cost for the first treatment arm
\code{meanCostB} mean cost for the second treatment arm
\code{meanOutcomeA} mean outome for the first treatment arm
\code{meanOutcomeB} mean outome for the second treatment arm
\code{seCostA} standard error of the cost of the first treatment arm
\code{seCostB} standard error of the cost of the second treatment arm
\code{seOutcomeA} standard error of the outcome of the first treatment arm
\code{seOutcomeB} standard error of the outcome of the second treatment arm
}
\description{
This is an internal function that takes as input a dataset with
4 columns \code{ID}, \code{RESPONSE}, \code{COST} and \code{TREATMENT} and extracts and generates
a list with the required information.
}
\details{
to be written
}
\author{
Amadou Gaye & Felix Achana
}
\keyword{internal}

