% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icer.R
\name{icer}
\alias{icer}
\title{Computes incremental cost-effectiveness ratio (ICER)}
\usage{
icer(data = NULL)
}
\arguments{
\item{data}{a dataset with 4 columns holding respectively, ID, treatment outcome,
treatment cost and treatment arm.}
}
\value{
a list which holds to elements: ICER value and other statistics for the given dataset
}
\description{
The function takes one dataset as input and the dataset
MUST have 4 columns \code{ID}, \code{RESPONSE}, \code{COST} and \code{TREATMENT}
which hold respectively unique subject identifiers, treatment outcome, treatment cost
and 2 treatment arms names.
}
\details{
to be written
}
\examples{
{

# load examples datasets
data(dataset1)

# ICER computation 
output <- icer(dataset1)

# display the ICER value
names(output)
output$ICER

# display the other stats 
output$Stats
}

}
\author{
Amadou Gaye & Felix Achana
}

