% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netbenef.R
\name{netbenef}
\alias{netbenef}
\title{Computes net benefit values across a range of wilingness-to-pay thresholds}
\usage{
netbenef(data = NULL, will2pay = NULL, CI = 0.95,
  treatResponse = "beneficial")
}
\arguments{
\item{data}{a dataset with 4 columns holding respectively, ID, treatment outcome,
treatment cost and treatment arm.}

\item{will2pay}{a numeric vector of willingness-to-pay thresholds.}

\item{CI}{confidence interval.}

\item{treatResponse}{a character, default is \code{beneficial} i.e. the treatment resulted in beneficial response;
otherwise \code{harmful}, the treatement resulted in harmful outcome}
}
\value{
a table with for columns: net benefit, standard error, upper and lower limit
}
\description{
The function takes one dataset and a vector of wilingness-to-pay values.
The dataset MUST have 4 columns \code{ID}, \code{RESPONSE}, \code{COST} and \code{TREATMENT}
which hold respectively unique subject identifiers, treatment outcome, treatment cost
and 2 treatment arms names. If no values are for wilingness-to-pay 101 default values ranging
from 0 to 500,000 are used to compute net befit.
}
\details{
to be written
}
\examples{
{

# load examples datasets
data(dataset1)

# nebefit computation using the default willingness-to-pay thresholds
NB <- netbenef(dataset1)

# display the top 5 records of the output table
NB[1:5,]

}

}
\author{
Amadou Gaye & Felix Achana
}

