% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icer.R
\name{icer}
\alias{icer}
\title{Computes incremental cost-effectiveness ratio (ICER)}
\usage{
icer(data1 = NULL, data2 = NULL)
}
\arguments{
\item{data1}{a dataset with 4 columns holding respectively, ID, treatment outcome,
treatment cost and treatment arm.}

\item{data1}{a dataset with 4 columns holding respectively, ID, treatment outcome,
treatment cost and treatment arm.}
}
\value{
a list of 2 lists which hold ICER value and other statistics for each dataset
}
\description{
The function takes at least one dataset as input and each dataset
MUST have 4 columns \code{ID}, \code{RESPONSE}, \code{COST} and \code{TREATMENT}
which hold respectively unique subject identifiers, treatment outcome, treatment cost
and 2 treatment arms names.
}
\examples{
{

# load examples datasets
load(dataset1)
load(dataset2)

# ICER computation 
output <- icer(dataset1, dataset2)

# display the ICER values for the two datasets
# first check what objects are in 'output' and then show results
names(output)
output$DATASET1$ICER
output$DATASET2$ICER

# we can also show all the results for a dataset
output$DATASET1
}

}
\author{
Amadou Gaye & Felix Achana
}

